# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Warehouse Management System Stub API
  version: 1.0.0
  description: |
    Versioned backend scaffold for web and mobile clients.
    All endpoints are stubs returning mock envelopes.
servers:
  - url: http://localhost:5000/api/v1
    description: Local backend
security:
  - bearerAuth: []
paths:
...
  /inventory/summary:
    get:
      tags: [Inventory]
      summary: Inventory summary
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          $ref: '#/components/responses/ListSuccessResponse'
...
  /inventory/adjustments:
    post:
      tags: [Inventory]
      summary: Create inventory adjustment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryAdjustmentRequest'
      responses:
        '201':
          $ref: '#/components/responses/SuccessResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
    PageSizeParam:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 500
  responses:
    SuccessResponse:
      description: Success envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessEnvelope'
    ListSuccessResponse:
      description: Success envelope with pagination metadata
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ListSuccessEnvelope'
...
  schemas:
    Meta:
      type: object
      properties:
        requestId:
          type: string
        tenantId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
      required: [requestId, tenantId, timestamp]
    Pagination:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, pageSize, total, totalPages]
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          additionalProperties: true
        meta:
          allOf:
            - $ref: '#/components/schemas/Meta'
      required: [success, data, meta]
    ListSuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            type: object
            additionalProperties: true
        meta:
          allOf:
            - $ref: '#/components/schemas/Meta'
            - type: object
              properties:
                pagination:
                  $ref: '#/components/schemas/Pagination'
      required: [success, data, meta]
...
    InventoryAdjustmentRequest:
      type: object
      properties:
        sku:
          type: string
        location:
          type: string
        qty:
          type: number
        reason:
          type: string
      required: [sku, location, qty, reason]
...